Standardize config keys, defaults and helpers
