+++ tracdap-libs/tracdap-lib-common/src/main/java/org/finos/tracdap/common/config/ConfigHelpers.java
@@ -110,4 +110,20 @@ public class ConfigHelpers {
         var message = String.format("Invalid boolean value for property [%s] in [%s]", key, context);
         throw new EStartup(message);
     }
+
+    public static String readOrDefault(String configValue, String defaultValue) {
+
+        if (configValue == null || configValue.isBlank())
+            return defaultValue;
+        else
+            return configValue;
+    }
+
+    public static int readOrDefault(int configValue, int defaultValue) {
+
+        if (configValue == 0)
+            return defaultValue;
+        else
+            return configValue;
+    }
 }
