+++ tracdap-services/tracdap-gateway/src/main/java/org/finos/tracdap/gateway/builders/ServiceInfo.java
@@ -22,7 +22,6 @@ import org.finos.tracdap.api.MetadataServiceProto;
 import org.finos.tracdap.api.OrchestratorServiceProto;
 import org.finos.tracdap.common.config.ConfigHelpers;
 import org.finos.tracdap.common.config.ConfigKeys;
-import org.finos.tracdap.common.config.ServiceProperties;
 import org.finos.tracdap.common.exception.ETracInternal;
 import org.finos.tracdap.config.PlatformConfig;
 import org.finos.tracdap.config.ServiceConfig;
@@ -170,14 +169,14 @@ public class ServiceInfo {
 
         var defaultServiceName = SERVICE_NAMES.get(serviceKey);
         this.serviceName = defaultServiceName != null
-                ? ConfigHelpers.readStringOrDefault(configContext, serviceProps, ServiceProperties.GATEWAY_ROUTE_NAME, defaultServiceName)
-                : ConfigHelpers.readString(configContext, serviceProps, ServiceProperties.GATEWAY_ROUTE_NAME);
+                ? ConfigHelpers.readStringOrDefault(configContext, serviceProps, ConfigKeys.GATEWAY_ROUTE_NAME, defaultServiceName)
+                : ConfigHelpers.readString(configContext, serviceProps, ConfigKeys.GATEWAY_ROUTE_NAME);
 
 
         var defaultHttpPrefix = SERVICE_PREFIX_DEFAULTS.get(serviceKey);
         var httpPrefix = defaultHttpPrefix != null
-                ? ConfigHelpers.readStringOrDefault(configContext, serviceProps, ServiceProperties.GATEWAY_ROUTE_PREFIX, defaultHttpPrefix)
-                : ConfigHelpers.readString(configContext, serviceProps, ServiceProperties.GATEWAY_ROUTE_PREFIX);
+                ? ConfigHelpers.readStringOrDefault(configContext, serviceProps, ConfigKeys.GATEWAY_ROUTE_PREFIX, defaultHttpPrefix)
+                : ConfigHelpers.readString(configContext, serviceProps, ConfigKeys.GATEWAY_ROUTE_PREFIX);
 
         if (restPrefix == null) {
             this.httpPrefix = httpPrefix;
